move current inliner heuristic to the InlinerHeuristics class
